+++ containers/Layout.tsx
@@ -1,3 +1,4 @@
+import useTawk from '@components/hooks/use-tawk'
 import Cookies from '@components/layouts/cookies'
 import Footer from '@components/layouts/footer'
 import ScrollTop from '@components/layouts/scroll-top'
@@ -21,9 +22,9 @@ import { FiFacebook, FiGithub, FiInstagram } from 'react-icons/fi'
 import { HiOutlineHeart } from 'react-icons/hi'
 import Cart from './cart/cart'
 import SearchbarBlock from './searchbar-block'
-import Chatwoot from '@components/Chatwoot'
 function ElevationScroll(props: any) {
   const { children, window } = props
+
   // Note that you normally won't need to set the window ref as useScrollTrigger
   // will default to window.
   // This is only being set here because the demo is in an iframe.
@@ -49,7 +50,7 @@ ElevationScroll.propTypes = {
 
 export default function Layout(props: any) {
   const { acceptedCookies, onAcceptCookies } = useAcceptCookies()
-
+  useTawk({ id: process.env.NEXT_PUBLIC_TAWK_ID })
   const socials = [
     {
       icon: <FiFacebook />,
@@ -106,7 +107,6 @@ export default function Layout(props: any) {
       <Container maxWidth="xl" sx={{ mt: 4, mb: 4 }}>
         {props.children}
       </Container>
-      <Chatwoot />
       <Cookies visible={!acceptedCookies} onClick={() => onAcceptCookies()} />
       <ScrollTop />
       <Footer
